Rename getToggles to fetchToggles
