+++ ws-nextjs-app/pages/[service]/onDemandTv/handleOnDemandTvRoute.ts
@@ -7,7 +7,7 @@ import getPageData from '#utilities/pageRequests/getPageData';
 import nodeLogger from '#lib/logger.node';
 import { ROUTING_INFORMATION } from '#app/lib/logger.const';
 import handleError from '#app/routes/utils/handleError';
-import fetchToggles from '#app/lib/utilities/getToggles';
+import fetchToggles from '#app/lib/utilities/fetchToggles';
 import isTest from '#app/lib/utilities/isTest';
 
 const logger = nodeLogger(__filename);
