+++ ws-nextjs-app/utilities/pageRequests/getPageData.test.ts
@@ -6,7 +6,7 @@ const agent = { cert: 'cert', ca: 'ca', key: 'key' };
 
 jest.mock('#utilities/getAgent', () => jest.fn(() => Promise.resolve(agent)));
 jest.mock('#app/routes/utils/fetchPageData');
-jest.mock('#app/lib/utilities/getToggles');
+jest.mock('#app/lib/utilities/fetchToggles');
 jest.mock('#app/routes/utils/fetchDataFromBFF', () => {
   const originalModule = jest.requireActual(
     '#app/routes/utils/fetchDataFromBFF',
